+++ src/lib/runner_test.rs
@@ -1148,20 +1148,19 @@ fn create_watch_task_with_makefile_and_all_object_options_and_cli_args() {
     make_command_line.push_str(" some_task 1 2 \"3 4\"");
 
     let args = task.args.unwrap();
-    assert_eq!(args.len(), 13);
+    assert_eq!(args.len(), 12);
     assert_eq!(args[0], "watch".to_string());
-    assert_eq!(args[1], "-q".to_string());
+    assert_eq!(args[1], "--why".to_string());
     assert_eq!(args[2], "--postpone".to_string());
     assert_eq!(args[3], "-i".to_string());
     assert_eq!(args[4], "tools/*".to_string());
     assert_eq!(args[5], "--no-gitignore".to_string());
-    assert_eq!(args[6], "--why".to_string());
-    assert_eq!(args[7], "-w".to_string());
-    assert_eq!(args[8], "dir1".to_string());
-    assert_eq!(args[9], "-w".to_string());
-    assert_eq!(args[10], "dir2".to_string());
-    assert_eq!(args[11], "-x".to_string());
-    assert_eq!(args[12], make_command_line.to_string());
+    assert_eq!(args[6], "-w".to_string());
+    assert_eq!(args[7], "dir1".to_string());
+    assert_eq!(args[8], "-w".to_string());
+    assert_eq!(args[9], "dir2".to_string());
+    assert_eq!(args[10], "-x".to_string());
+    assert_eq!(args[11], make_command_line.to_string());
 }
 
 #[test]
@@ -1211,20 +1210,19 @@ fn create_watch_task_with_makefile_and_all_object_options() {
     make_command_line.push_str(" some_task");
 
     let args = task.args.unwrap();
-    assert_eq!(args.len(), 13);
+    assert_eq!(args.len(), 12);
     assert_eq!(args[0], "watch".to_string());
-    assert_eq!(args[1], "-q".to_string());
+    assert_eq!(args[1], "--why".to_string());
     assert_eq!(args[2], "--postpone".to_string());
     assert_eq!(args[3], "-i".to_string());
     assert_eq!(args[4], "tools/*".to_string());
     assert_eq!(args[5], "--no-gitignore".to_string());
-    assert_eq!(args[6], "--why".to_string());
-    assert_eq!(args[7], "-w".to_string());
-    assert_eq!(args[8], "dir1".to_string());
-    assert_eq!(args[9], "-w".to_string());
-    assert_eq!(args[10], "dir2".to_string());
-    assert_eq!(args[11], "-x".to_string());
-    assert_eq!(args[12], make_command_line.to_string());
+    assert_eq!(args[6], "-w".to_string());
+    assert_eq!(args[7], "dir1".to_string());
+    assert_eq!(args[8], "-w".to_string());
+    assert_eq!(args[9], "dir2".to_string());
+    assert_eq!(args[10], "-x".to_string());
+    assert_eq!(args[11], make_command_line.to_string());
 }
 
 #[test]