+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.36.9
 
 * Fix: '--hide-uninteresting'  cli flag #859 (thanks @xxchan)
+* Fix: cargo-watch option 'why' should remove '--q' #849
 
 ### v0.36.8 (2023-05-27)